Rename to camelCase more consistently
