+++ __tests__/components/ui/codeBlock.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { CodeBlock } from '../../../components/ui/code-block';
-import { CopyButton } from '../../../components/ui/copy-button';
+import { CodeBlock } from '../../../components/ui/codeBlock';
+import { CopyButton } from '../../../components/ui/copyButton';
 
 // Mock CopyButton component
-jest.mock('../../../components/ui/copy-button', () => ({
+jest.mock('../../../components/ui/copyButton', () => ({
   CopyButton: jest.fn(({ content }) => (
     <button data-testid="mock-copy-button" data-content={content}>
       Copy
