+++ __tests__/components/ui/themeToggle.test.tsx
@@ -1,5 +1,5 @@
 import { render, fireEvent, screen } from '@testing-library/react';
-import { ThemeToggle } from '@/components/ui/theme-toggle';
+import { ThemeToggle } from '@/components/ui/themeToggle';
 import { useTheme } from 'next-themes';
 
 // Mock next-themes
