+++ components/features/bookmarks/bookmark-card.client.tsx
@@ -22,8 +22,8 @@
  */
 
 import { ExternalLink as ExternalLinkIcon } from 'lucide-react';
-import { ExternalLink } from '../../ui/external-link';
-import { LogoImage } from '../../ui/logo-image';
+import { ExternalLink } from '../../ui/externalLink';
+import { LogoImage } from '../../ui/logoImage';
 import type { Bookmark } from '../../../types/bookmark';
 import { normalizeDomain } from '../../../lib/logo-fetcher';
