+++ components/features/education/certification-card.client.tsx
@@ -20,7 +20,7 @@
 
 import { ExternalLink as ExternalLinkIcon } from 'lucide-react';
 import { LogoImage } from '../../../components/ui';
-import { ExternalLink } from '../../../components/ui/external-link';
+import { ExternalLink } from '../../ui/externalLink';
 import type { Certification } from '../../../types/education';
 
 /**
