+++ components/ui/financialMetrics.tsx
